+++ GUI/coregui/Views/ImportDataWidgets/RealDataSelectorToolBar.cpp
@@ -14,7 +14,7 @@
 //
 // ************************************************************************** //
 
-#include "ImportDataToolBar.h"
+#include "RealDataSelectorToolBar.h"
 #include "ComboProperty.h"
 #include "DetectorItems.h"
 #include "IDetector2D.h"
