+++ GUI/coregui/Views/ImportDataWidgets/RealDataSelectorWidget.cpp
@@ -19,7 +19,7 @@
 #include "RealDataPropertiesWidget.h"
 #include "RealDataModel.h"
 #include "minisplitter.h"
-#include "ImportDataToolBar.h"
+#include "RealDataSelectorToolBar.h"
 #include <QVBoxLayout>
 #include <QItemSelectionModel>
